# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bible Atlas API
  description: The Bible Atlas HTTP API, generated from the Rust that serves it.
  version: 0.8.0
paths:
...
  /api/catechism/item/{id}:
    get:
      tags:
      - catechism
      summary: |-
        `GET /api/catechism/item/{id}` (Batch F, "item fetch by id" --
        requirement 3). Unknown id -> 404 `not_found`, same precedent
        `places::place` already set for an exact-identifier lookup (not a ref
        with its own "out of canon but still valid shape" middle ground). Each
        proof verse resolves to its own full KJV text (`CatechismProofVerse`'s
        own doc comment); a verse id that fails to resolve (should never happen
        -- `validate::run_catechism` already guarantees every curated verse
        exists in the compiled KJV text) is skipped rather than panicking, same
        ruling-4 soft-fail policy `reading::verse`'s own cross-ref preview
        lookup already follows.
      description: |-
        OVERLAY-1 Task 2: proof-verse text now comes from `graph.verse_text_of`
        (a real, on-demand graph query), not the retired `AtlasData.verses` --
        this handler picks up a second extractor, `State<Arc<GraphService>>`,
        the same combined-state pattern `reading::verse` already uses.
      operationId: catechism_item
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatechismItem'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
  /api/catechism/{sref}:
    get:
      tags:
      - catechism
      summary: |-
        `GET /api/catechism/{sref}` (Batch F, "verse -> citing catechism items
        lookup" -- requirement 3): `sref` must parse as exactly a
        `ScriptureRef::Verse` or `ScriptureRef::Passage`, mirroring
        `reading::xrefs`'s own accepted-shapes precedent exactly (a bare book or
        chapter ref has no defined "member verses" to aggregate over, so both 400
        as `bad_ref`). ruling-3-policy: an sref with no citing catechism items at
        all is NOT an error -- 200 with an empty list, same "gracefully empty,
        never a 404" policy `xrefs`/`chapter`/`scene_scripture` already follow.
        Business logic (the union-across-member-verses aggregation) lives in
        `atlas_core::catechism::items_for_span`, reached here via
        `AtlasData::catechism_items_for_span` -- this handler is pure
        response-shape assembly, same as every other handler in this crate.
      description: |-
        PROV-1 FIX ROUND 1 (review M-3): now takes `State<Arc<GraphService>>`
        too, purely to attribute each row. That is the SECOND extractor the
        review thought would make this half "a genuinely larger change" -- it is
        not: `AppState` already implements `FromRef<AppState>` for
        `Arc<GraphService>`, and six handlers in this crate already take both.
        The aggregation itself is untouched; `AtlasData` is still where the
        business logic lives.
      operationId: catechism_for_span
      parameters:
      - name: sref
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatechismRef'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
  /api/chapter/{cref}:
    get:
      tags:
      - reading
      summary: |-
        `GET /api/chapter/{cref}`. `cref` must parse as exactly a
        `ScriptureRef::Chapter` (book + chapter, e.g. `EXO.14`) — a book-only or
        verse/passage-shaped path segment is the wrong shape for this endpoint
        and 400s as `bad_ref`, same as an unparseable one. The optional
        `?translation=kjv` query param (ruling 5, M1 is KJV-only) is never
        extracted, so its presence or absence cannot affect this handler at all.
      description: |-
        ruling-3-policy: once `cref` parses as a `Chapter`, an out-of-range
        chapter number (or a book with no known chapters in this atlas) is NOT an
        error — the verse-count bound comes from `canon.books[].chapters`, and an
        unknown/short chapter just yields `verse_count = 0`, i.e. a 200 response
        with an empty `verses` list. Same rationale as `scene_scripture`: a
        reader showing "no verses in this chapter" is a meaningful response, not
        a failure.

        Batch M-A (brief requirement 5, "re-implement the OLD /api/chapter
        handler as a VIEW over the window query"): the verse TEXT below now
        comes from `GraphState::chapter_span` + `GraphState::window` -- the SAME
        windowed reading-order query `GET /api/text?scope=chapter` calls --
        instead of `data.verses.get(key)`. The verse-count bound and the
        out-of-canon policy above are UNCHANGED and still sourced from
        `AtlasData`; headings moved to `graph.heading_index` in M-C2, and
        OVERLAY-1 Task 5 moved the PLACE-MENTION half onto the port too
        (`graph.scene_source(&data)`'s own `places_for_verse`/`place`, the
        materialised-from-the-graph successors of the deleted
        `AtlasData::places_for_verse`/`place_by_id` -- identical ids in
        identical order, see those methods' own doc comments). THIS endpoint
        (the reader's own chapter view) was untouched by Batch M-B's own
        event-world migration;
        only `/api/narrative/event/{id}` (see that handler's own doc comment)
        and the generic `/api/node`/`/edges` endpoints move to the graph this
        batch. The WIRE SHAPE is byte-for-byte identical -- proven by
        `tests/graph_equivalence.rs`'s own all-1,189-chapters comparison -- so
        every existing caller of this endpoint (the reader's chapter view/
        mini-reader/split view, `ChapterNode`, `PlaceCard`'s hover verse text,
        `PassageBlock`, `PopoverSectionProviders`) now serves from the graph
        with NO client-side change and no reader-visible behavior change.
      operationId: chapter
      parameters:
      - name: cref
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
...
  /api/scene/scripture:
    get:
      tags:
      - map
      summary: '`GET /api/scene/scripture?ref=`.'
      description: |-
        ruling-3-policy: a `ref` that fails to *parse* (unknown book code, zero
        chapter/verse, empty segment, inverted range — i.e. `ScriptureRef::parse`
        returns `Err`) is a structurally bad ref and always 400s as `bad_ref`,
        same as a missing `ref` param (treated as parsing the empty string, which
        also fails to parse). A `ref` that *parses* but names coordinates outside
        the loaded canon (e.g. a chapter number past the end of the book) is
        deliberately NOT an error: `compose_scripture_scene` only ever matches it
        against verses that actually exist on events/places, so an out-of-canon
        ref naturally composes an empty-but-valid scene, exactly mirroring
        ruling 2's "don't reject out-of-span time windows" for the sibling
        endpoint. This needs no extra bounds-checking code — it falls out of not
        adding any.

        OVERLAY-1 Task 5: composes from `graph.scene_source(&data)`, exactly as
        `scene_time` above does -- see that handler's own doc comment.
      operationId: scene_scripture
      parameters:
      - name: ref
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scene'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
...
components:
  schemas:
...
    CatechismItem:
      type: object
      description: |-
        `GET /api/catechism/item/{id}`'s own wire shape -- `part_title` alongside
        the item's own fields so the client never needs a second fetch to show
        "Baptism" as this item's own chief-part context. `where_written` is
        omitted (not `null`) when absent -- same conditional-presence wire
        convention `History.blurb`/`established`/`destroyed` already use.
      required:
      - id
      - name
      - part_title
      - explanation_heading
      - explanation
      - verses
      properties:
        explanation:
          type: string
        explanation_heading:
          type: string
        id:
          type: string
        name:
          type: string
        part_title:
          type: string
        text:
          type:
          - string
          - 'null'
        verses:
          type: array
          items:
            $ref: '#/components/schemas/CatechismProofVerse'
        where_written:
          type:
          - string
          - 'null'
      additionalProperties: false
    CatechismProofVerse:
      type: object
      description: |-
        Batch F: one resolved proof verse -- `vref` + its own FULL KJV text
        (design-direction.md's house rendering, per requirement 4: "THE
        SCRIPTURES -- the item's proof verses... full verse text per house
        rendering" -- not a truncated preview the way `CrossRef.preview` is).
        Batch F2: `question` (omitted when absent, same convention as
        `CatechismRef.question`) names which question this proof verse came
        from, when it came from one -- `None` for Luther's own item-level
        embedded citations. This is requirement 4's own "if cheap, highlight/
        deep-link the question context": a small caption next to the verse in
        THE SCRIPTURES, cheap because it needs no new fetch or scroll machinery,
        just this one extra field already available at merge time.
      required:
      - vref
      - text
      properties:
        question:
          type:
          - string
          - 'null'
        text:
          type: string
        vref:
          type: string
      additionalProperties: false
    CatechismRef:
      type: object
      description: |-
        Batch F ("the small catechism"), extended Batch F2 (question-level
        citations): one catechism item citing a verse/span, as embedded in
        `VerseDetail.catechism` AND returned (as a list) by `GET
        /api/catechism/{sref}` -- deliberately lean (id + display name + an
        optional question title, no preview text unlike `CrossRef`):
        requirement 4's own UI description lists citing items (now
        question-aware, "<Item> — <Question title>") as plain named entries,
        nothing more, and the full item content is a separate fetch (`GET
        /api/catechism/item/{id}`, see `CatechismItem` below), made only once
        the user actually opens one. `question` is omitted from the wire
        entirely (not null) when this hit came from Luther's own item-level
        embedded citation rather than a question -- same conditional-presence
        convention `CatechismItem.where_written` below and `wire/places.rs`'s own
        `History.blurb` already use.
      required:
      - id
      - name
      - provenance
      properties:
        id:
          type: string
        name:
          type: string
        provenance:
          type: array
          items:
            type: string
          description: |-
            PROV-1 FIX ROUND 1 (review M-3): the `catechism` family's own
            distinct provenance set, so a PASSAGE node's THE SMALL CATECHISM
            section gets a "?" like a verse's. Same additive element-level move
            and same set-not-single-id reasoning as `CrossRef.provenance` --
            see that field's doc comment.

            The review judged deferring this half defensible because
            `catechism_for_span` took only `State<Arc<AtlasData>>`. Measured, it
            is not: `AppState` implements `FromRef` for `Arc<GraphService>` too,
            and six handlers in this crate already take BOTH extractors
            (`reading::verse` among them). So the second half was one extractor
            away, not a larger change, and it ships here.

            This family is the genuinely MULTI-sourced one --
            `{concord-sc-overlap, curated-catechism}`, pinned -- which is exactly
            why the field is a `Vec` on both endpoints.

            GRANULARITY (fix round 2, review M-NEW-1): SECTION-level, carried on
            the element because the endpoint is a bare array. Every element of
            one response receives the identical set. Not per-row, and not to be
            described as per-row -- see `CrossRef.provenance` for the full
            statement and for why per-row is unavailable at these call sites.
        question:
          type:
          - string
          - 'null'
      additionalProperties: false
    Chapter:
      type: object
      required:
      - ref
      - book
      - chapter
      - verses
      properties:
        book:
          type: string
        chapter:
          type: integer
          format: int32
          minimum: 0
        ref:
          type: string
        verses:
          type: array
          items:
            $ref: '#/components/schemas/Verse'
      additionalProperties: false
...
    Scene:
      type: object
      required:
      - mode
      - places
      - quiet_places
      - arrows
      - narratives
      properties:
        arrows:
          type: array
          items:
            $ref: '#/components/schemas/SceneArrow'
        mode:
          $ref: '#/components/schemas/SceneMode'
        narratives:
          type: array
          items:
            $ref: '#/components/schemas/SceneNarrative'
        places:
          type: array
          items:
            $ref: '#/components/schemas/ScenePlace'
        quiet_places:
          type: array
          items:
            $ref: '#/components/schemas/QuietPlace'
          description: |-
            Batch E2 (the ever-present graph, user direction 2026-08-19: "all of
            the cities in our graph are available in any timerange rather than
            just loading those which are biblically active at the time"): every
            event-bearing place NOT in `places` for this window -- period
            relevance is now expressed by GLOW (whether a place is in `places`),
            not existence. QUIET-1: for every time-mode window, `places` union
            `quiet_places` equals the FULL event-bearing place set (fixed
            cardinality, derived from the data -- see
            `AtlasData::event_bearing_place_ids`), and the two are disjoint.
            ALWAYS an array (never omitted) on both modes -- but always EMPTY for
            scripture-mode scenes, which never gain quiet places at all (chosen
            over an `Option`/omitted-key representation for wire regularity with
            `places` itself, which is likewise never optional; a client reading
            `quiet_places` needs no null-check either way).
        ref:
          type:
          - string
          - 'null'
        window:
          oneOf:
          - $ref: '#/components/schemas/TimeRange'
          - type: 'null'
      additionalProperties: false
    SceneArrow:
      type: object
      required:
      - narrative
      - color
      - from_place
      - to_place
      - from_event
      - to_event
      - order
      properties:
        color:
          type: string
        from_event:
          type: string
        from_place:
          type: string
        narrative:
          type: string
        order:
          type: integer
          format: int32
          minimum: 0
        to_event:
          type: string
        to_place:
          type: string
      additionalProperties: false
    SceneEvent:
      type: object
      required:
      - id
      - label
      - when
      - verse_groups
      properties:
        id:
          type: string
        label:
          type: string
        verse_groups:
          type: array
          items:
            $ref: '#/components/schemas/VerseGroup'
        when:
          $ref: '#/components/schemas/TimeRange'
      additionalProperties: false
    SceneMode:
      type: string
      description: |-
        Which question a scene answers: what was happening in a span of years, or
        where a passage happens. Every scene is composed by exactly one of
        `scene::compose_time_scene`/`compose_scripture_scene`, so the pair is
        closed by construction.
      enum:
      - time
      - scripture
    SceneNarrative:
      type: object
      required:
      - id
      - name
      - color
      - legs_in_scene
      properties:
        color:
          type: string
        id:
          type: string
        legs_in_scene:
          type: integer
          format: int32
          minimum: 0
        name:
          type: string
      additionalProperties: false
    ScenePlace:
      type: object
      required:
      - id
      - name
      - display_name
      - lat
      - lon
      - brightness
      - events
      properties:
        brightness:
          type: integer
          format: int32
          minimum: 0
        display_name:
          type: string
          description: |-
            Batch E: the period-true name resolved for this scene's own window
            (`crate::history::resolve_display_name`) -- `name`'s original,
            always-present value unchanged; this is the ADDITIONAL field the
            client renders on map labels / hover card title / arrow-tip text.
            Equal to `name` whenever this place has no curated history, or none
            of its curated ranges intersects the window.
        events:
          type: array
          items:
            $ref: '#/components/schemas/SceneEvent'
        existence_from:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            Batch H (existence gating, deferred from E2): the curated
            established/destroyed bounds this place's name resolves against
            (`crate::history::resolve_existence`) -- plain years, never the full
            established/destroyed `PlaceDateClaim` (verses/note; that richer
            shape is `/api/place/{id}`'s own `wire::History`, not a per-marker wire
            cost). Both omitted (not `null` -- see `skip_serializing_if`) when
            this place has no curated history at all, or a curated history with
            neither claim -- the client's own gate never fires in that case
            (`existence_gates_label`'s own "always labels" reading). The map
            renders every place regardless of these bounds (a lit place is only
            ever here because a real event justifies it THIS window -- see
            `scene::lit_places`); it's the client's own label -- never the
            marker/dot -- that these bounds gate, and only in time mode (there is
            no window to test outside-ness against in scripture mode).
        existence_to:
          type:
          - integer
          - 'null'
          format: int32
        id:
          type: string
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        merged_ids:
          type: array
          items:
            type: string
          description: |-
            Batch HOTFIX-2 (same-place dedupe): ids of every OTHER compiled place
            record merged into this one at load time (`crate::merge::MERGE_PAIRS`)
            -- e.g. `["hazor_545"]` on the place carrying id `hazor-1`. Wire
            traceability only (no client behavior reads this today): the brief's
            own "note the absorbed record's id in the wire for traceability" --
            so a future investigation into "why does hazor-1 carry events from
            two different upstream lineages" doesn't have to go spelunking
            through source data by hand. Empty (and omitted -- `skip_serializing_if`,
            matching `existence_from`/`_to`'s own lean-wire convention) for the
            overwhelming majority of places, which were never part of a merge.
        name:
          type: string
      additionalProperties: false
...
    TimeRange:
      type: object
      required:
      - from_year
      - to_year
      properties:
        from_year:
          type: integer
          format: int32
        to_year:
          type: integer
          format: int32
...
    Verse:
      type: object
      required:
      - verse
      - text
      - places
      - persons
      - xref_count
      - words_of_christ
      properties:
        heading:
          oneOf:
          - $ref: '#/components/schemas/Heading'
            description: |-
              Batch T requirement 5: this verse's own pericope heading, when it is
              a heading ANCHOR (`AtlasData::heading_for_verse`) -- omitted (not
              null), matching `Polity.transition`/`.fall`'s own conditional-
              presence convention, since the overwhelming majority of verses never
              anchor a heading at all.
          - type: 'null'
        persons:
          type: array
          items:
            $ref: '#/components/schemas/PersonRef'
          description: |-
            M-D3 (owner ruling U5): every person the graph's own `mentions`
            relation attests at this verse's own locus, in wire (row-insertion)
            order -- see `PersonRef`'s own doc comment for the full
            mentions-attested-only reasoning. Always present, possibly empty
            (most verses mention zero persons by name) -- SAME conditional-
            presence-lives-client-side convention `places` immediately above
            already establishes for this exact shape of field.
        places:
          type: array
          items:
            $ref: '#/components/schemas/PlaceRef'
          description: |-
            Batch R requirement 5: every curated place whose OWN `verse_links`
            names this verse, in place-id order (stable, not meaningful --
            there is no "primary mention" concept). Always present, possibly
            empty (most verses geocode to zero places) -- conditional presence
            lives on the CLIENT side (a mini-reader renders no hoverable span at
            all when this is empty), not as an omitted wire key, matching every
            other "always an array" field in this app's own wire (e.g.
            `Scene.quiet_places`).
        text:
          type: string
        verse:
          type: integer
          format: int32
          minimum: 0
        words_of_christ:
          type: array
          items:
            $ref: '#/components/schemas/WordsOfChristSpan'
          description: |-
            Batch RED-1: this verse's own aligned sub-verse red-letter spans
            (`GraphService.red_letter_spans`, precomputed once -- see that
            field's own doc comment for why this ONE companion is not
            graph-derivable), in ascending order. Always present, possibly
            empty -- SAME "always an array" convention `places`/`persons`
            above already establish (most verses carry none). KJV-ONLY by
            construction (decision 5: sub-verse precision is KJV-specific;
            this endpoint has never served any other translation -- ruling 5,
            "M1 is KJV-only").
        xref_count:
          type: integer
          description: |-
            Batch M-D2 (owner's cross-reference superscript directive, closed on
            the graph platform): this verse's own `cites` edge-summary COUNT at
            its TextUnit locus -- read straight off THE PORT (`GraphQuery::
            edge_summary`, design doc §5), the SAME generic query the `/api/node/
            {id}` handler itself calls (`graph::node_card`) -- never a
            second, bespoke tally. This is a genuinely NEW field (no bespoke
            predecessor existed), sourced NATIVELY from the generic contract from
            day one: the reader's own superscript markers ARE a query of the
            graph, per the design's own P1/P4 laws, just batched onto the
            EXISTING per-chapter response (one HTTP round trip for a whole
            chapter) rather than issuing 20-30 individual `GET /api/node/{id}`
            calls per chapter, which `client/IExplorableClient.cs`'s own concrete
            implementation is what a SINGLE node's lookup uses (see that file's
            own doc comment + `GraphExplorableClientTests` for the direct,
            unbatched proof the same contract is correct in isolation). Always
            present (0 for the overwhelming majority of verses, never omitted --
            the client's own superscript decision is `count == 0` -> nothing,
            never a missing-field special case).
          minimum: 0
      additionalProperties: false
...